Changes made:
- Re-assigned a few a accelerators to allocate CTRL+7/8/9/0 for dynamically changing the file's encoding
- Fixed two (dynamic) cast warnings for ScintillaWin.cxx

+++ scintilla/win32/ScintillaWin.cxx
@@ -2075,12 +2075,12 @@ public:
 			}
 
 			const size_t lenOut = ::WideCharToMultiByte(cp, 0,
-				&utf16Folded[0], lenFlat,
+				&utf16Folded[0], static_cast<int>(lenFlat),
 				NULL, 0, NULL, 0);
 
 			if (lenOut < sizeFolded) {
 				::WideCharToMultiByte(cp, 0,
-					&utf16Folded[0], lenFlat,
+					&utf16Folded[0], static_cast<int>(lenFlat),
 					folded, static_cast<int>(lenOut), NULL, 0);
 				return lenOut;
 			} else {
